Fixes (hopefully last) tests
